+++ sdk/resourcemanager/botservice/armbotservice/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armbotservice
 
 const (
 	moduleName    = "armbotservice"
-	moduleVersion = "v0.2.0"
+	moduleVersion = "v0.2.1"
 )
 
 type ChannelName string