+++ sdk/resourcemanager/dataprotection/armdataprotection/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armdataprotection
 
 const (
 	moduleName    = "armdataprotection"
-	moduleVersion = "v0.2.0"
+	moduleVersion = "v0.2.1"
 )
 
 type AbsoluteMarker string