+++ sdk/resourcemanager/kubernetesconfiguration/armkubernetesconfiguration/ze_generated_example_clusterextensiontypes_client_test.go
@@ -25,8 +25,8 @@ func ExampleClusterExtensionTypesClient_List() {
 	ctx := context.Background()
 	client := armkubernetesconfiguration.NewClusterExtensionTypesClient("<subscription-id>", cred, nil)
 	pager := client.List("<resource-group-name>",
-		armkubernetesconfiguration.Enum0("Microsoft.ContainerService"),
-		armkubernetesconfiguration.Enum1("managedClusters"),
+		armkubernetesconfiguration.ExtensionsClusterRp("Microsoft.ContainerService"),
+		armkubernetesconfiguration.ExtensionsClusterResourceName("managedClusters"),
 		"<cluster-name>",
 		nil)
 	for {