+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/ze_generated_example_recoverypointsrecommendedformove_client_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup"
 )
 
-// x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-10-01/examples/AzureIaasVm/RecoveryPointsRecommendedForMove_List.json
+// x-ms-original-file: specification/recoveryservicesbackup/resource-manager/Microsoft.RecoveryServices/stable/2021-12-01/examples/AzureIaasVm/RecoveryPointsRecommendedForMove_List.json
 func ExampleRecoveryPointsRecommendedForMoveClient_List() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {