+++ sdk/resourcemanager/relay/armrelay/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armrelay
 
 const (
 	moduleName    = "armrelay"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.3.1"
 )
 
 type AccessRights string