+++ sdk/resourcemanager/resources/armsubscriptions/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armsubscriptions
 
 const (
 	moduleName    = "armsubscriptions"
-	moduleVersion = "v0.2.0"
+	moduleVersion = "v0.2.1"
 )
 
 // LocationType - The location type.