+++ sdk/resourcemanager/testbase/armtestbase/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armtestbase
 
 const (
 	moduleName    = "armtestbase"
-	moduleVersion = "v0.2.0"
+	moduleVersion = "v0.2.1"
 )
 
 // Action - The action of the command.